roster animap sprite fix LivestockManager wool and double quality buddywnd grp fix dev log fix new mountmeter hide option minimize all windows new pbot commands widget Unknown position operation?

+++ src/haven/livestock/Animal.java
@@ -17,7 +17,7 @@ import java.util.List;
 import java.util.Map;
 
 
-public abstract class Animal extends HashMap<String, Integer> {
+public abstract class Animal extends HashMap<String, Double> {
     private final static int COL_SPACE;
     public long gobid;
     public long wndid;
